Use `u16` instead of `usize` for line width
Using an `usize` for the line width causes problems for the
optimal-fit wrapping algorithm. The problem is that the algorithm
works with integer values formed by

    (line_width - text_width)**2

When `line_width` is near `usize::max_value()`, this computation will
overflow an `usize`. By limiting the line width to an `u16`, we can do
the internal computations with `u64` and avoid the overflows.

+++ benches/linear.rs
@@ -9,7 +9,7 @@ use criterion::{criterion_group, criterion_main};
 
 use lipsum::lipsum_words_from_seed;
 
-const LINE_LENGTH: usize = 60;
+const LINE_LENGTH: u16 = 60;
 
 /// Generate a lorem ipsum text with the given number of characters.
 fn lorem_ipsum(length: usize) -> String {
